fix: hide all internal classes (#12)
* fix: hide all internal classes

* expose KotlinDataFetcher
