+++ src/main/kotlin/com/expedia/graphql/schema/exceptions/CouldNotGetNameOfEnumException.kt
@@ -1,3 +1,6 @@
 package com.expedia.graphql.schema.exceptions
 
-class CouldNotGetNameOfEnumException : RuntimeException()
+import kotlin.reflect.KClass
+
+class CouldNotGetNameOfEnumException(kclass: KClass<*>)
+    : RuntimeException("Could not get the enum name of the KClass $kclass")
