+++ src/main/kotlin/com/expedia/graphql/schema/exceptions/TypeNotSupportedException.kt
@@ -1,4 +1,4 @@
 package com.expedia.graphql.schema.exceptions
 
-class TypeNotSupportedException(typeName: String, packageName: String) :
-    RuntimeException("Cannot convert $typeName since it is outside the supported package $packageName")
+class TypeNotSupportedException(typeName: String, packageName: String)
+    : RuntimeException("Cannot convert $typeName since it is outside the supported package $packageName")
