+++ src/test/kotlin/com/expedia/graphql/conflicts/GeoType.kt
@@ -1,4 +1,4 @@
-package com.expedia.graphql.schema.conflicts
+package com.expedia.graphql.conflicts
 
 // This should conflict with the model in SchemaGeneratorTest.kt
 enum class GeoType {
